+++ apps/server/src/modules/room/api/service/index.ts
@@ -1,0 +1,1 @@
+export * from './room-permission.service';
