--- apps/server/src/modules/room/service/index.ts
@@ -1,1 +1,0 @@
-export * from './room.helper';
